fix(billing): scope event preaggregation to temporal workflow
